api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
web3==7.7.0
eth-account==0.13.4
pydantic==2.10.4

flask==3.0.3
python-dotenv==1.0.1